--- .bazelrc
@@ -1,8 +1,0 @@
-build --verbose_failures
-
-# This is so we don't use sandboxed execution. Sandboxed execution
-# runs stuff in a container, and since Travis already runs its script
-# in a container (unless you require sudo in your .travis.yml) this
-# fails to run tests.
-build --spawn_strategy=standalone --genrule_strategy=standalone
-test --test_strategy=standalone
