+++ .github/workflows/platformio.yml
@@ -1,0 +1,28 @@
+on: [push, pull_request]
+
+jobs:
+  test:
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [ubuntu-latest]
+
+    steps:
+    - uses: actions/checkout@v1
+    - name: Set up Python
+      uses: actions/setup-python@v1
+    - name: Install cppcheck
+      run: sudo apt-get install -y cppcheck
+    - name: Install PlatformIO
+      run: |
+        python -m pip install --upgrade pip
+        pip install -U platformio
+        platformio upgrade --dev          
+        platformio update
+        platformio --version
+    - name: Tests on the native platform
+      run: platformio test -e desktop
+    - name: Build examples
+      run: platformio ci examples/ --lib=. --board=uno --board=nodemcuv2
+    - name: Static Code Analysis
+      run: cppcheck src/
