+++ .gitignore
@@ -1,5 +1,5 @@
-# Ignore the bazel created symlinks.
-/bazel-*
-
 # Ignore vim swap files.
 *.swp
+
+/.pio
+/.vscode
