--- .travis.yml
@@ -1,27 +1,0 @@
-dist: trusty
-
-notifications:
-  email:
-    - mum4k-ci@elohim.sk
-
-addons:
-  apt:
-    sources:
-      - ubuntu-toolchain-r-test
-    packages:
-      - pkg-config
-      - python
-      - python-pip
-      - unzip
-      - wget
-      - zip
-
-before_install:
-  - wget https://github.com/bazelbuild/bazel/releases/download/0.23.1/bazel_0.23.1-linux-x86_64.deb
-  - sha256sum -c bazel/bazel_0.23.1-linux-x86_64.deb.sha256
-  - sudo dpkg --force-all -i bazel_0.23.1-linux-x86_64.deb
-  - sudo /usr/bin/pip install platformio
-
-script:
-  - bazel build //...
-  - bazel test --test_output all //...
