--- arduino/arduino_impl.cc
@@ -1,34 +1,0 @@
-// Copyright 2017 Google Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#include <Arduino_impl.h>
-
-#include <Arduino.h>
-#include <Arduino_interface.h>
-
-namespace arduino {
-
-ArduinoImpl::~ArduinoImpl() {}
-
-int ArduinoImpl::DigitalRead(uint8_t pin) const { return digitalRead(pin); }
-
-void ArduinoImpl::DigitalWrite(uint8_t pin, uint8_t value) const {
-  return digitalWrite(pin, value);
-}
-
-unsigned long ArduinoImpl::Millis(void) const { return millis(); }
-
-void ArduinoImpl::Delay(unsigned long ms) const { delay(ms); }
-
-} // namespace arduino
