--- bazel/bazel_0.23.1-linux-x86_64.deb.sha256
@@ -1,1 +1,0 @@
-62d7fc733cb64c8bcedec4374e674ffacdc6616584d913fe84b97753c5e0863e  bazel_0.23.1-linux-x86_64.deb
