+++ platformio.ini
@@ -1,0 +1,30 @@
+; PlatformIO Project Configuration File
+;
+;   Build options: build flags, source filter
+;   Upload options: custom upload port, speed and extra flags
+;   Library options: dependencies, extra library storages
+;   Advanced options: extra scripting
+;
+; Please visit documentation for the other options and examples
+; https://docs.platformio.org/page/projectconf.html
+
+[env:desktop]
+platform = native
+build_flags = 
+	-std=gnu++11
+	-pthread
+lib_deps = 
+	googletest
+lib_compat_mode = off 
+
+[env:uno]
+platform = atmelavr
+board = uno
+framework = arduino
+lib_ignore = googletest
+
+[env:nodemcuv2]
+platform = espressif8266
+board = nodemcuv2
+framework = arduino
+lib_ignore = googletest
