+++ todo/software_serial_impl.cc
